+++ lib/renderers/server.js
