+++ lib/server.js
@@ -1,6 +1,6 @@
 import express  from 'express';
 import  config from './config.js';
-import  serverRender  from './serverRender.js';
+import  serverRender  from '../lib/renderers/server.js';
 
 const app = express();
