+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-    entry: ['babel-polyfill','./lib/components/index.js'],
+    entry: ['babel-polyfill','./lib/renderers/dom.js'],
     output: {
         path: path.resolve(__dirname,'public'),
         filename: 'bundle.js'
